% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffiXcan.R
\name{affiXcanGReX}
\alias{affiXcanGReX}
\title{Compute a GReX from variables and their coefficients for a set of genes}
\usage{
affiXcanGReX(affiXcanTraining, pcs, cores)
}
\arguments{
\item{affiXcanTraining}{The returning object from affiXcanTrain()}

\item{pcs}{A list, which is the returning object from affiXcanPcs()}

\item{cores}{An integer >0; if cores=1 processes will not be parallelized}
}
\value{
A SummarizedExperiment object containing the imputed GReX values
}
\description{
Compute a GReX from variables and their coefficients for a set of genes
}
\examples{
\dontrun{
trainingTbaPaths <- system.file("extdata","training.tba.toydata.rds",
package="AffiXcan")

data(exprMatrix)
data(regionAssoc)
data(trainingCovariates)

assay <- "values"

training <- affiXcanTrain(exprMatrix=exprMatrix, assay=assay,
tbaPaths=trainingTbaPaths, regionAssoc=regionAssoc, cov=trainingCovariates,
varExplained=80, scale=TRUE, cores=1)

testingTbaPaths <- system.file("extdata","testing.tba.toydata.rds",
package="AffiXcan")

pcs <- affiXcanPcs(tbaPaths=testingTbaPaths, affiXcanTraining=training,
scale=TRUE, cores=2)

bs <- training$bs

exprmatrix <- affiXcanGReX(affiXcanTraining=training, cores=1)
}
}
