% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffiXcan.R
\name{computePca}
\alias{computePca}
\title{Perform a PCA on a matrix where columns are variables}
\usage{
computePca(data, varExplained, scale)
}
\arguments{
\item{data}{A matrix containing the TBA values for a certain genomic region;
columns are PWMs, rows are individuals IIDs}

\item{varExplained}{An integer between 0 and 100; varExplained=80 means that
the principal components selected to fit the models must explain at least
80 percent of variation of TBA values}

\item{scale}{A logical; if scale=FALSE the TBA values will be only centered,
not scaled before performing PCA}
}
\value{
A list containing two objects:

    eigenvectors: a matrix containing eigenvectors for those principal
    components selected according to the param varExplained

    pcs: a matrix containing the principal components values selected
    according to the param varExplained
}
\description{
Perform a PCA on a matrix where columns are variables
}
\examples{
tbaMatrixMAE <- readRDS(system.file("extdata","training.tba.toydata.rds",
package="AffiXcan"))

tbaMatrix <- MultiAssayExperiment::experiments(tbaMatrixMAE)

tba <- tbaMatrix$ENSG00000256377.1

pca <- computePca(data=tba, varExplained=80, scale=TRUE)
}
