% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffiXcan.R
\name{subsetKFold}
\alias{subsetKFold}
\title{Split the numbers from 1 to n in k equally-sized lists}
\usage{
subsetKFold(k, n)
}
\arguments{
\item{k}{An integer. The number of groups in which the first n natural
numbers are to be splitted}

\item{n}{An integer. Defines the interval 1..n}
}
\value{
sampGroups: a list of lists; the first natural n numbers equally
distributed across k lists
}
\description{
Split the numbers from 1 to n in k equally-sized lists
}
\examples{
sampGroups <- subsetKFold(k=5, n=23)
}
