% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{trainingCovariates}
\alias{trainingCovariates}
\title{Covariates of the population structure for 230 individuals}
\format{An object of class \code{"data.frame"}}
\usage{
data(trainingCovariates)
}
\description{
Toy data used in examples to describe affiXcanTrain() function;
Columns are the first three principal components of the population structure
Rows are individials
}
\examples{
data(trainingCovariates)
head(trainingCovariates)
}
\keyword{datasets}
