% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffiXcan.R
\name{affiXcanImpute}
\alias{affiXcanImpute}
\title{Impute a GReX for each gene for whom a model was generated}
\usage{
affiXcanImpute(tbaPaths, pca, bs, scale, cores)
}
\arguments{
\item{tbaPaths}{A vector of strings, which are the paths to
MultiAssayExperiment RDS files containing the tba values}

\item{pca}{A list, which is the returningObject$pca from affiXcanTrain()}

\item{bs}{A list, which is the returningObject$bs from affiXcanTrain()}

\item{scale}{A logical; if scale=FALSE the TBA values will be only centered,
not scaled before performing PCA}

\item{cores}{An integer >0; if cores=1 processes will not be parallelized}
}
\value{
A SummarizedExperiment object containing imputed GReX values
}
\description{
Impute a GReX for each gene for whom a model was generated
}
\examples{
trainingTbaPaths <- system.file("extdata","training.tba.toydata.rds",
package="AffiXcan")

data(exprMatrix)
data(regionAssoc)
data(trainingCovariates)

assay <- "values"

training <- affiXcanTrain(exprMatrix=exprMatrix, assay=assay,
tbaPaths=trainingTbaPaths, regionAssoc=regionAssoc, cov=trainingCovariates,
varExplained=80, scale=TRUE, cores=1)

testingTbaPaths <- system.file("extdata","testing.tba.toydata.rds",
package="AffiXcan")

exprmatrix <- affiXcanImpute(tbaPaths=testingTbaPaths, pca=training$pca,
bs=training$bs, scale=TRUE, cores=1)
}
