% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AffiXcan.R
\name{affiXcanPcs}
\alias{affiXcanPcs}
\title{Compute PCs in MultiAssayExperiment objects using eigenvectors given by user}
\usage{
affiXcanPcs(tbaPaths, pca, scale, cores)
}
\arguments{
\item{tbaPaths}{A vector of strings, which are the paths to
MultiAssayExperiment RDS files containing the tba values}

\item{pca}{A list, which is the returningObject$pca from affiXcanTrain()}

\item{scale}{A logical; if scale=FALSE the TBA values will be only centered,
not scaled before performing PCA}

\item{cores}{An integer >0; if cores=1 processes will not be parallelized}
}
\value{
A list of matrices containing the principal components values of TBA
for each region; each object of the list is named after the
MultiAssayExperiment object from which it derives
}
\description{
Compute PCs in MultiAssayExperiment objects using eigenvectors given by user
}
\examples{
trainingTbaPaths <- system.file("extdata","training.tba.toydata.rds",
package="AffiXcan")

data(exprMatrix)
data(regionAssoc)
data(trainingCovariates)

assay <- "values"

training <- affiXcanTrain(exprMatrix=exprMatrix, assay=assay,
tbaPaths=trainingTbaPaths, regionAssoc=regionAssoc, cov=trainingCovariates,
varExplained=80, scale=TRUE, cores=1)

testingTbaPaths <- system.file("extdata","testing.tba.toydata.rds",
package="AffiXcan")

pcs <- affiXcanPcs(tbaPaths=testingTbaPaths, pca=training$pca, scale=TRUE,
cores=1)
}
