% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{exprMatrix}
\alias{exprMatrix}
\title{Expression data of four genes for 230 individuals}
\format{An object of class \code{"SummarizedExperiment"}.}
\usage{
data(exprMatrix)
}
\description{
Toy dataset used in examples to describe affiXcanTrain() function
}
\examples{
library(SummarizedExperiment)
data(exprMatrix)
toyExpressionMatrix <- assays(exprMatrix)$values
}
\keyword{datasets}
