% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetsDoc.R
\docType{data}
\name{regionAssoc}
\alias{regionAssoc}
\title{Associations between regulatory regions and expressed genes}
\format{An object of class \code{"data.frame"}}
\usage{
data(regionAssoc)
}
\description{
Toy data used in examples to describe affiXcanTrain() and affiXcanImpute() 
functions; it has two columns: for every EXPRESSED_REGION are listed the
associated REGULATORY_REGION(s)
}
\examples{
data(regionAssoc)
head(regionAssoc)
}
\keyword{datasets}
